Source::get_line_text no longer trims end whitespace
This was causing crashes using byte spans when byte spans start or end in whitespace after a line. When I write the code I did not expect get_line_text to trim whitespace.

I edited uses of get_line_text to do the trimming separately.

Technically this is a breaking change.